# Coucher.WebApi/swagger-live.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Coucher.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/exercises": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exercises/{id}/participants": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/participants/{participantId}/role": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRole"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRole"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRole"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/manager": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseParticipant"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/sections": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSection"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSection"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSection"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/influencers": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseInfluencer"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseInfluencer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseInfluencer"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/contacts": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseUnitContactRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseUnitContactRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseUnitContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/contacts/{contactId}": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseUnitContactRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseUnitContactRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseUnitContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUnitContact"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/exercises/{id}/archive": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/unarchive": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/participants/{participantId}": {
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/exercises/sections/{sectionLinkId}": {
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "sectionLinkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/exercises/influencers/{influencerLinkId}": {
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "influencerLinkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateExerciseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "traineeUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "trainerUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "compressionFactor": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "influencerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "sectionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "managerUserId": {
            "type": "string",
            "nullable": true
          },
          "traineeUnitContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseUnitContactRequest"
            },
            "nullable": true
          },
          "trainerUnitContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseUnitContactRequest"
            },
            "nullable": true
          },
          "additionalParticipants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseParticipantRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Exercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "traineeUnitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "trainerUnitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "statusId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "compressionFactor": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdateTime": {
            "type": "string",
            "format": "date-time"
          },
          "completionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "archiveTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "archivedByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "traineeUnit": {
            "$ref": "#/components/schemas/Unit"
          },
          "trainerUnit": {
            "$ref": "#/components/schemas/Unit"
          },
          "status": {
            "$ref": "#/components/schemas/ClosedListItem"
          },
          "createdByUser": {
            "$ref": "#/components/schemas/UserProfile"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseParticipant"
            },
            "nullable": true
          },
          "unitContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseUnitContact"
            },
            "nullable": true
          },
          "influencers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseInfluencer"
            },
            "nullable": true
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSection"
            },
            "nullable": true
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseTask"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExerciseParticipant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "exerciseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/ExerciseRole"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "exercise": {
            "$ref": "#/components/schemas/Exercise"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "additionalProperties": false
      },
      "ExerciseParticipantRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/ExerciseRole"
          }
        },
        "additionalProperties": false
      },
      "ExerciseRole": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ExerciseSection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "exerciseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sectionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "exercise": {
            "$ref": "#/components/schemas/Exercise"
          },
          "section": {
            "$ref": "#/components/schemas/ClosedListItem"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateExerciseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "statusId": {
            "type": "string",
            "format": "uuid"
          },
          "traineeUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "trainerUnitId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "SessionId": {
        "type": "apiKey",
        "name": "session-id",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "SessionId": [ ]
    }
  ]
}